+++ CHANGELOG.rst
@@ -16,6 +16,13 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.11
+----
+2025-03-24
+
+- Optimise ``file_name`` provider. Add ``prefix`` argument.
+- Add ``file_path`` and ``dir_path`` providers.
+
 0.10.5
 ------
 2024-11-23